+++ src/main/kotlin/platform/mixin/handlers/injectionPoint/FieldInjectionPoint.kt
@@ -21,9 +21,9 @@
 package com.demonwav.mcdev.platform.mixin.handlers.injectionPoint
 
 import com.demonwav.mcdev.platform.mixin.reference.MixinSelector
+import com.demonwav.mcdev.platform.mixin.util.MemberInfo
 import com.demonwav.mcdev.platform.mixin.util.fakeResolve
 import com.demonwav.mcdev.platform.mixin.util.findOrConstructSourceField
-import com.demonwav.mcdev.util.MemberReference
 import com.demonwav.mcdev.util.constantValue
 import com.intellij.codeInsight.completion.JavaLookupElementBuilder
 import com.intellij.codeInsight.lookup.LookupElementBuilder
@@ -97,7 +97,7 @@ class FieldInjectionPoint : QualifiedInjectionPoint<PsiField>() {
         mode: CollectVisitor.Mode,
     ): CollectVisitor<PsiField>? {
         if (mode == CollectVisitor.Mode.COMPLETION) {
-            return MyCollectVisitor(mode, at.project, MemberReference(""), -1, null, 8)
+            return MyCollectVisitor(mode, at.project, MemberInfo(""), -1, null, 8)
         }
         val opcode = (at.findDeclaredAttributeValue("opcode")?.constantValue as? Int)
             ?.takeIf { it in Const.VALID_OPCODES } ?: -1
