+++ src/main/kotlin/util/MemberReference.kt
@@ -3,7 +3,7 @@
  *
  * https://mcdev.io/
  *
- * Copyright (C) 2025 minecraft-dev
+ * Copyright (C) 2026 minecraft-dev
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published
@@ -20,26 +20,19 @@
 
 package com.demonwav.mcdev.util
 
-import com.demonwav.mcdev.platform.mixin.reference.MixinSelector
-import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiField
 import com.intellij.psi.PsiMethod
-import java.io.Serializable
 import org.objectweb.asm.Type
 
 /**
- * Represents a reference to a class member (a method or a field). It may
- * resolve to multiple members if [matchAllNames] or [matchAllDescs] is set or if the member is
- * not full qualified.
+ * Represents a reference to a class member (a method or a field).
  */
 data class MemberReference(
     val name: String,
     val descriptor: String? = null,
     override val owner: String? = null,
-    val matchAllNames: Boolean = false,
-    val matchAllDescs: Boolean = false,
-) : Serializable, MixinSelector {
+) : MemberMatcher {
 
     init {
         assert(owner?.contains('/') != true)
@@ -61,7 +54,6 @@ data class MemberReference(
 
     override val methodDescriptor = descriptor?.takeIf { it.contains("(") }
     override val fieldDescriptor = descriptor?.takeUnless { it.contains("(") }
-    override val displayName = name
 
     val presentableText: String get() = buildString {
         if (owner != null) {
@@ -77,7 +69,7 @@ data class MemberReference(
     }
 
     override fun canEverMatch(name: String): Boolean {
-        return matchAllNames || this.name == name
+        return this.name == name
     }
 
     private fun matchOwner(clazz: String): Boolean {
@@ -87,90 +79,39 @@ data class MemberReference(
 
     override fun matchField(owner: String, name: String, desc: String): Boolean {
         assert(!owner.contains('.'))
-        return (this.matchAllNames || this.name == name) &&
+        return this.name == name &&
             matchOwner(owner) &&
             (this.descriptor == null || this.descriptor == desc)
     }
 
     override fun matchMethod(owner: String, name: String, desc: String): Boolean {
         assert(!owner.contains('.'))
-        return (this.matchAllNames || this.name == name) &&
+        return this.name == name &&
             matchOwner(owner) &&
             (this.descriptor == null || this.descriptor == desc)
     }
 
-    companion object {
-        fun parse(value: String): MemberReference? {
-            val reference = value.replace(" ", "")
-            val owner: String?
+    fun toMixinString() = buildString {
+        if (owner != null) {
+            append('L').append(owner.replace('.', '/')).append(';')
+        }
 
-            var pos = reference.lastIndexOf('.')
-            if (pos != -1) {
-                // Everything before the dot is the qualifier/owner
-                owner = reference.substring(0, pos).replace('/', '.')
-            } else {
-                pos = reference.indexOf(';')
-                if (pos != -1 && reference.startsWith('L')) {
-                    val internalOwner = reference.substring(1, pos)
-                    if (!StringUtil.isJavaIdentifier(internalOwner.replace('/', '_'))) {
-                        // Invalid: Qualifier should only contain slashes
-                        return null
-                    }
+        append(name)
 
-                    owner = internalOwner.replace('/', '.')
-
-                    // if owner is all there is to the selector, match anything with the owner
-                    if (pos == reference.length - 1) {
-                        return MemberReference("", null, owner, matchAllNames = true, matchAllDescs = true)
-                    }
-                } else {
-                    // No owner/qualifier specified
-                    pos = -1
-                    owner = null
-                }
+        descriptor?.let { descriptor ->
+            if (!descriptor.startsWith('(')) {
+                // Field descriptor
+                append(':')
             }
 
-            val descriptor: String?
-            val name: String
-            val matchAllNames = reference.getOrNull(pos + 1) == '*'
-            val matchAllDescs: Boolean
-
-            // Find descriptor separator
-            val methodDescPos = reference.indexOf('(', pos + 1)
-            if (methodDescPos != -1) {
-                // Method descriptor
-                descriptor = reference.substring(methodDescPos)
-                name = reference.substring(pos + 1, methodDescPos)
-                matchAllDescs = false
-            } else {
-                val fieldDescPos = reference.indexOf(':', pos + 1)
-                if (fieldDescPos != -1) {
-                    descriptor = reference.substring(fieldDescPos + 1)
-                    name = reference.substring(pos + 1, fieldDescPos)
-                    matchAllDescs = false
-                } else {
-                    descriptor = null
-                    matchAllDescs = reference.endsWith('*')
-                    name = if (matchAllDescs) {
-                        reference.substring(pos + 1, reference.lastIndex)
-                    } else {
-                        reference.substring(pos + 1)
-                    }
-                }
-            }
-
-            if (!matchAllNames && !StringUtil.isJavaIdentifier(name) && name != "<init>" && name != "<clinit>") {
-                return null
-            }
-
-            return MemberReference(if (matchAllNames) "*" else name, descriptor, owner, matchAllNames, matchAllDescs)
+            append(descriptor)
         }
     }
 }
 
 // Class
 
-fun PsiClass.findMethods(member: MixinSelector, checkBases: Boolean = false): Sequence<PsiMethod> {
+fun PsiClass.findMethods(member: MemberMatcher, checkBases: Boolean = false): Sequence<PsiMethod> {
     val methods = if (checkBases) {
         allMethods.asSequence()
     } else {
@@ -179,7 +120,7 @@ fun PsiClass.findMethods(member: MixinSelector, checkBases: Boolean = false): Se
     return methods.filter { member.matchMethod(it, this) }
 }
 
-fun PsiClass.findField(selector: MixinSelector, checkBases: Boolean = false): PsiField? {
+fun PsiClass.findField(selector: MemberMatcher, checkBases: Boolean = false): PsiField? {
     val fields = if (checkBases) {
         allFields.toList()
     } else {
